Edited on 27 - 2

+++ print_functions.c
@@ -20,7 +20,7 @@ int print_char(va_list list)
 
 int print_str(va_list list)
 {
-unsigned int i;
+unsigned long long i;
 
 char *str = va_arg(list, char*);
